+++ src/modules/complianceengine/tests/RegexTest.cpp
@@ -21,7 +21,7 @@ TEST_F(RegexTest, NoMatch)
     bool result = regex_search(input, match, r);
     EXPECT_FALSE(result);
     ASSERT_TRUE(match.ready());
-    EXPECT_EQ(match.size(), 0);
+    EXPECT_EQ(match.size(), (size_t)0);
 }
 
 TEST_F(RegexTest, Match)
@@ -35,9 +35,9 @@ TEST_F(RegexTest, Match)
     bool result = regex_search(input, match, r);
     EXPECT_TRUE(result);
     ASSERT_TRUE(match.ready());
-    ASSERT_EQ(match.size(), 1);
+    ASSERT_EQ(match.size(), (size_t)1);
     EXPECT_EQ(match[0].matched, true);
-    EXPECT_EQ(match[0].length(), 4);
+    EXPECT_EQ(match[0].length(), (ssize_t)4);
 }
 
 TEST_F(RegexTest, MatchWithSubMatches_1)
@@ -49,11 +49,11 @@ TEST_F(RegexTest, MatchWithSubMatches_1)
     bool result = regex_search(input, match, r);
     EXPECT_TRUE(result);
     ASSERT_TRUE(match.ready());
-    ASSERT_EQ(match.size(), 2);
+    ASSERT_EQ(match.size(), (size_t)2);
     EXPECT_EQ(match[0].matched, true);
-    EXPECT_EQ(match[0].length(), std::strlen("test"));
+    EXPECT_EQ(match[0].length(), (ssize_t)std::strlen("test"));
     EXPECT_EQ(match[1].matched, true);
-    EXPECT_EQ(match[1].length(), std::strlen("test"));
+    EXPECT_EQ(match[1].length(), (ssize_t)std::strlen("test"));
 }
 
 TEST_F(RegexTest, MatchWithSubMatches_2)
@@ -65,13 +65,13 @@ TEST_F(RegexTest, MatchWithSubMatches_2)
     bool result = regex_search(input, match, r);
     EXPECT_TRUE(result);
     ASSERT_TRUE(match.ready());
-    ASSERT_EQ(match.size(), 3);
+    ASSERT_EQ(match.size(), (size_t)3);
     EXPECT_EQ(match[0].matched, true);
-    EXPECT_EQ(match[0].length(), std::strlen("test string"));
+    EXPECT_EQ(match[0].length(), (ssize_t)std::strlen("test string"));
     EXPECT_EQ(match[1].matched, true);
-    EXPECT_EQ(match[1].length(), std::strlen("test"));
+    EXPECT_EQ(match[1].length(), (ssize_t)std::strlen("test"));
     EXPECT_EQ(match[2].matched, true);
-    EXPECT_EQ(match[2].length(), std::strlen("string"));
+    EXPECT_EQ(match[2].length(), (ssize_t)std::strlen("string"));
 }
 
 TEST_F(RegexTest, MatchWithSubMatches_3)
@@ -83,15 +83,15 @@ TEST_F(RegexTest, MatchWithSubMatches_3)
     bool result = regex_search(input, match, r);
     EXPECT_TRUE(result);
     ASSERT_TRUE(match.ready());
-    ASSERT_EQ(match.size(), 4);
+    ASSERT_EQ(match.size(), (size_t)4);
     EXPECT_EQ(match[0].matched, true);
-    EXPECT_EQ(match[0].length(), std::strlen("test string"));
+    EXPECT_EQ(match[0].length(), (ssize_t)std::strlen("test string"));
     EXPECT_EQ(match[1].matched, true);
-    EXPECT_EQ(match[1].length(), std::strlen("test string"));
+    EXPECT_EQ(match[1].length(), (ssize_t)std::strlen("test string"));
     EXPECT_EQ(match[2].matched, true);
-    EXPECT_EQ(match[2].length(), std::strlen("test"));
+    EXPECT_EQ(match[2].length(), (ssize_t)std::strlen("test"));
     EXPECT_EQ(match[3].matched, true);
-    EXPECT_EQ(match[3].length(), std::strlen("string"));
+    EXPECT_EQ(match[3].length(), (ssize_t)std::strlen("string"));
     EXPECT_EQ(match[100].matched, false);
     EXPECT_EQ(match[100].length(), 0);
 }
@@ -135,7 +135,7 @@ TEST_F(RegexTest, RegexMatch_1)
     bool result = regex_match(target, match, r);
     EXPECT_FALSE(result);
     EXPECT_TRUE(match.ready());
-    EXPECT_EQ(match.size(), 0);
+    EXPECT_EQ(match.size(), (size_t)0);
     EXPECT_TRUE(match.empty());
 }
 
@@ -148,7 +148,7 @@ TEST_F(RegexTest, RegexMatch_2)
     bool result = regex_match(target, match, r);
     EXPECT_FALSE(result);
     EXPECT_TRUE(match.ready());
-    EXPECT_EQ(match.size(), 0);
+    EXPECT_EQ(match.size(), (size_t)0);
     EXPECT_TRUE(match.empty());
 }
